Rename all private modules with leading underscore
Resolves type checking issues from previous commit.
