+++ setup.cfg
@@ -61,7 +61,7 @@ ignore_errors = True
 
 # Disable strict optional for telegram objects with class methods
 # We don't want to clutter the code with 'if self.bot is None: raise RuntimeError()'
-[mypy-telegram.callbackquery,telegram.chat,telegram.message,telegram.user,telegram.files.*,telegram.inline.inlinequery,telegram.payment.precheckoutquery,telegram.payment.shippingquery,telegram.passport.passportdata,telegram.passport.credentials,telegram.passport.passportfile,telegram.ext.filters]
+[mypy-telegram._callbackquery,telegram._chat,telegram._message,telegram._user,telegram.files.*,telegram.inline._inlinequery,telegram.payment._precheckoutquery,telegram.payment._shippingquery,telegram.passport._passportdata,telegram.passport._credentials,telegram.passport._passportfile,telegram.ext.filters]
 strict_optional = False
 
 # type hinting for asyncio in webhookhandler is a bit tricky because it depends on the OS
