+++ telegram/files/_file.py
@@ -25,7 +25,7 @@ from os.path import basename
 from typing import IO, TYPE_CHECKING, Any, Optional, Union
 
 from telegram import TelegramObject
-from telegram.passport.credentials import decrypt
+from telegram.passport._credentials import decrypt
 from telegram.utils.files import is_local_file
 
 if TYPE_CHECKING:
