+++ telegram/payment/_shippingquery.py
